+++ src/CanvasRenderingContext2d.cc
@@ -21,8 +21,6 @@
 
 using namespace v8;
 
-Nan::Persistent<FunctionTemplate> Context2d::constructor;
-
 /*
  * Rectangle arg assertions.
  */
@@ -83,22 +81,20 @@ inline static bool checkArgs(const Nan::FunctionCallbackInfo<Value> &info, doubl
   return areArgsValid;
 }
 
-Nan::Persistent<Function> Context2d::_DOMMatrix;
-Nan::Persistent<Function> Context2d::_parseFont;
-
 /*
  * Initialize Context2d.
  */
 
 void
-Context2d::Initialize(Nan::ADDON_REGISTER_FUNCTION_ARGS_TYPE target) {
+Context2d::Initialize(Nan::ADDON_REGISTER_FUNCTION_ARGS_TYPE target, AddonData* addon_data) {
   Nan::HandleScope scope;
 
+  Local<External> data_holder = Nan::New<External>(addon_data);
   // Constructor
-  Local<FunctionTemplate> ctor = Nan::New<FunctionTemplate>(Context2d::New);
-  constructor.Reset(ctor);
-  ctor->InstanceTemplate()->SetInternalFieldCount(1);
+  Local<FunctionTemplate> ctor = Nan::New<FunctionTemplate>(Context2d::New, data_holder);
+  ctor->InstanceTemplate()->SetInternalFieldCount(2);
   ctor->SetClassName(Nan::New("CanvasRenderingContext2D").ToLocalChecked());
+  addon_data->context2d_ctor_tpl.Reset(ctor);
 
   // Prototype
   Local<ObjectTemplate> proto = ctor->PrototypeTemplate();
@@ -167,7 +163,7 @@ Context2d::Initialize(Nan::ADDON_REGISTER_FUNCTION_ARGS_TYPE target) {
   SetProtoAccessor(proto, Nan::New("textAlign").ToLocalChecked(), GetTextAlign, SetTextAlign, ctor);
   Local<Context> ctx = Nan::GetCurrentContext();
   Nan::Set(target, Nan::New("CanvasRenderingContext2d").ToLocalChecked(), ctor->GetFunction(ctx).ToLocalChecked());
-  Nan::Set(target, Nan::New("CanvasRenderingContext2dInit").ToLocalChecked(), Nan::New<Function>(SaveExternalModules));
+  Nan::Set(target, Nan::New("CanvasRenderingContext2dInit").ToLocalChecked(), Nan::New<Function>(SaveExternalModules, data_holder));
 }
 
 /*
@@ -707,8 +703,10 @@ NAN_METHOD(Context2d::New) {
 
   if (!info[0]->IsObject())
     return Nan::ThrowTypeError("Canvas expected");
+  AddonData *addon_data = reinterpret_cast<AddonData*>(info.Data().As<External>()->Value());
   Local<Object> obj = Nan::To<Object>(info[0]).ToLocalChecked();
-  if (!Nan::New(Canvas::constructor)->HasInstance(obj))
+  Local<FunctionTemplate> canvas_ctor = Nan::New(addon_data->canvas_ctor_tpl);
+  if (!canvas_ctor->HasInstance(obj))
     return Nan::ThrowTypeError("Canvas expected");
   Canvas *canvas = Nan::ObjectWrap::Unwrap<Canvas>(obj);
 
@@ -743,6 +741,7 @@ NAN_METHOD(Context2d::New) {
   Context2d *context = new Context2d(canvas);
 
   context->Wrap(info.This());
+  info.This()->SetInternalField(1, info.Data());
   info.GetReturnValue().Set(info.This());
 }
 
@@ -751,8 +750,9 @@ NAN_METHOD(Context2d::New) {
  */
 
 NAN_METHOD(Context2d::SaveExternalModules) {
-  _DOMMatrix.Reset(Nan::To<Function>(info[0]).ToLocalChecked());
-  _parseFont.Reset(Nan::To<Function>(info[1]).ToLocalChecked());
+  AddonData *addon_data = reinterpret_cast<AddonData*>(info.Data().As<External>()->Value());
+  addon_data->context2d_dom_matrix.Reset(Nan::To<Function>(info[0]).ToLocalChecked());
+  addon_data->context2d_parse_font.Reset(Nan::To<Function>(info[1]).ToLocalChecked());
 }
 
 /*
@@ -806,7 +806,9 @@ NAN_METHOD(Context2d::PutImageData) {
   if (!info[0]->IsObject())
     return Nan::ThrowTypeError("ImageData expected");
   Local<Object> obj = Nan::To<Object>(info[0]).ToLocalChecked();
-  if (!Nan::New(ImageData::constructor)->HasInstance(obj))
+  AddonData *addon_data = get_data_from_if1(info.This());
+  Local<FunctionTemplate> image_data_ctor_tpl = Nan::New(addon_data->image_data_ctor_tpl);
+  if (!image_data_ctor_tpl->HasInstance(obj))
     return Nan::ThrowTypeError("ImageData expected");
 
   Context2d *context = Nan::ObjectWrap::Unwrap<Context2d>(info.This());
@@ -1154,7 +1156,8 @@ NAN_METHOD(Context2d::GetImageData) {
   Local<Int32> shHandle = Nan::New(sh);
   Local<Value> argv[argc] = { dataArray, swHandle, shHandle };
 
-  Local<Function> ctor = Nan::GetFunction(Nan::New(ImageData::constructor)).ToLocalChecked();
+  AddonData *addon_data = get_data_from_if1(info.This());
+  Local<Function> ctor = Nan::GetFunction(Nan::New(addon_data->image_data_ctor_tpl)).ToLocalChecked();
   Local<Object> instance = Nan::NewInstance(ctor, argc, argv).ToLocalChecked();
 
   info.GetReturnValue().Set(instance);
@@ -1195,7 +1198,8 @@ NAN_METHOD(Context2d::CreateImageData){
   const int argc = 3;
   Local<Value> argv[argc] = { arr, Nan::New(width), Nan::New(height) };
 
-  Local<Function> ctor = Nan::GetFunction(Nan::New(ImageData::constructor)).ToLocalChecked();
+  AddonData *addon_data = get_data_from_if1(info.This());
+  Local<Function> ctor = Nan::GetFunction(Nan::New(addon_data->image_data_ctor_tpl)).ToLocalChecked();
   Local<Object> instance = Nan::NewInstance(ctor, argc, argv).ToLocalChecked();
 
   info.GetReturnValue().Set(instance);
@@ -1249,10 +1253,13 @@ NAN_METHOD(Context2d::DrawImage) {
 
   cairo_surface_t *surface;
 
+  AddonData *addon_data = get_data_from_if1(info.This());
   Local<Object> obj = Nan::To<Object>(info[0]).ToLocalChecked();
+  Local<FunctionTemplate> canvas_ctor_tpl = Nan::New(addon_data->canvas_ctor_tpl);
+  Local<FunctionTemplate> image_ctor_tpl = Nan::New(addon_data->image_ctor_tpl);
 
   // Image
-  if (Nan::New(Image::constructor)->HasInstance(obj)) {
+  if (image_ctor_tpl->HasInstance(obj)) {
     Image *img = Nan::ObjectWrap::Unwrap<Image>(obj);
     if (!img->isComplete()) {
       return Nan::ThrowError("Image given has not completed loading");
@@ -1262,7 +1269,7 @@ NAN_METHOD(Context2d::DrawImage) {
     surface = img->surface();
 
   // Canvas
-  } else if (Nan::New(Canvas::constructor)->HasInstance(obj)) {
+  } else if (canvas_ctor_tpl->HasInstance(obj)) {
     Canvas *canvas = Nan::ObjectWrap::Unwrap<Canvas>(obj);
     source_w = sw = canvas->getWidth();
     source_h = sh = canvas->getHeight();
@@ -1778,7 +1785,7 @@ NAN_SETTER(Context2d::SetQuality) {
  */
 
 Local<Object>
-get_current_transform(Context2d *context) {
+get_current_transform(Context2d *context, AddonData *addon_data) {
   Isolate *iso = Isolate::GetCurrent();
 
   Local<Float64Array> arr = Float64Array::New(ArrayBuffer::New(iso, 48), 0, 6);
@@ -1794,7 +1801,8 @@ get_current_transform(Context2d *context) {
 
   const int argc = 1;
   Local<Value> argv[argc] = { arr };
-  return Nan::NewInstance(context->_DOMMatrix.Get(iso), argc, argv).ToLocalChecked();
+  Local<Function> _DOMMatrix = Nan::New(addon_data->context2d_dom_matrix);
+  return Nan::NewInstance(_DOMMatrix, argc, argv).ToLocalChecked();
 }
 
 /*
@@ -1818,8 +1826,9 @@ void parse_matrix_from_object(cairo_matrix_t &matrix, Local<Object> mat) {
  */
 
 NAN_GETTER(Context2d::GetCurrentTransform) {
+  AddonData *addon_data = get_data_from_if1(info.This());
   Context2d *context = Nan::ObjectWrap::Unwrap<Context2d>(info.This());
-  Local<Object> instance = get_current_transform(context);
+  Local<Object> instance = get_current_transform(context, addon_data);
 
   info.GetReturnValue().Set(instance);
 }
@@ -1829,12 +1838,14 @@ NAN_GETTER(Context2d::GetCurrentTransform) {
  */
 
 NAN_SETTER(Context2d::SetCurrentTransform) {
+  AddonData *addon_data = get_data_from_if1(info.This());
   Context2d *context = Nan::ObjectWrap::Unwrap<Context2d>(info.This());
   Local<Context> ctx = Nan::GetCurrentContext();
   Local<Object> mat = Nan::To<Object>(value).ToLocalChecked();
+  Local<Function> _DOMMatrix = Nan::New(addon_data->context2d_dom_matrix);
 
 #if NODE_MAJOR_VERSION >= 8
-  if (!mat->InstanceOf(ctx, _DOMMatrix.Get(Isolate::GetCurrent())).ToChecked()) {
+  if (!mat->InstanceOf(ctx, _DOMMatrix).ToChecked()) {
     return Nan::ThrowTypeError("Expected DOMMatrix");
   }
 #endif
@@ -1868,6 +1879,7 @@ NAN_GETTER(Context2d::GetFillStyle) {
 
 NAN_SETTER(Context2d::SetFillStyle) {
   Context2d *context = Nan::ObjectWrap::Unwrap<Context2d>(info.This());
+  AddonData *addon_data = get_data_from_if1(info.This());
 
   if (value->IsString()) {
     MaybeLocal<String> mstr = Nan::To<String>(value);
@@ -1876,12 +1888,14 @@ NAN_SETTER(Context2d::SetFillStyle) {
     context->_fillStyle.Reset();
     context->_setFillColor(str);
   } else if (value->IsObject()) {
+    Local<FunctionTemplate> gradient_ctor_tpl = Nan::New(addon_data->gradient_ctor_tpl);
+    Local<FunctionTemplate> pattern_ctor_tpl = Nan::New(addon_data->pattern_ctor_tpl);
     Local<Object> obj = Nan::To<Object>(value).ToLocalChecked();
-    if (Nan::New(Gradient::constructor)->HasInstance(obj)) {
+    if (gradient_ctor_tpl->HasInstance(obj)) {
       context->_fillStyle.Reset(value);
       Gradient *grad = Nan::ObjectWrap::Unwrap<Gradient>(obj);
       context->state->fillGradient = grad->pattern();
-    } else if (Nan::New(Pattern::constructor)->HasInstance(obj)) {
+    } else if (pattern_ctor_tpl->HasInstance(obj)) {
       context->_fillStyle.Reset(value);
       Pattern *pattern = Nan::ObjectWrap::Unwrap<Pattern>(obj);
       context->state->fillPattern = pattern->pattern();
@@ -1911,6 +1925,7 @@ NAN_GETTER(Context2d::GetStrokeStyle) {
 
 NAN_SETTER(Context2d::SetStrokeStyle) {
   Context2d *context = Nan::ObjectWrap::Unwrap<Context2d>(info.This());
+  AddonData *addon_data = get_data_from_if1(info.This());
 
   if (value->IsString()) {
     MaybeLocal<String> mstr = Nan::To<String>(value);
@@ -1920,11 +1935,13 @@ NAN_SETTER(Context2d::SetStrokeStyle) {
     context->_setStrokeColor(str);
   } else if (value->IsObject()) {
     Local<Object> obj = Nan::To<Object>(value).ToLocalChecked();
-    if (Nan::New(Gradient::constructor)->HasInstance(obj)) {
+    Local<FunctionTemplate> pattern_ctor_tpl = Nan::New(addon_data->pattern_ctor_tpl);
+    Local<FunctionTemplate> gradient_ctor_tpl = Nan::New(addon_data->gradient_ctor_tpl);
+    if (gradient_ctor_tpl->HasInstance(obj)) {
       context->_strokeStyle.Reset(value);
       Gradient *grad = Nan::ObjectWrap::Unwrap<Gradient>(obj);
       context->state->strokeGradient = grad->pattern();
-    } else if (Nan::New(Pattern::constructor)->HasInstance(obj)) {
+    } else if (pattern_ctor_tpl->HasInstance(obj)) {
       context->_strokeStyle.Reset(value);
       Pattern *pattern = Nan::ObjectWrap::Unwrap<Pattern>(obj);
       context->state->strokePattern = pattern->pattern();
@@ -2127,6 +2144,7 @@ Local<Value> Context2d::_getStrokeColor() {
 }
 
 NAN_METHOD(Context2d::CreatePattern) {
+  AddonData *addon_data = get_data_from_if1(info.This());
   Local<Value> image = info[0];
   Local<Value> repetition = info[1];
 
@@ -2136,7 +2154,7 @@ NAN_METHOD(Context2d::CreatePattern) {
   const int argc = 2;
   Local<Value> argv[argc] = { image, repetition };
 
-  Local<Function> ctor = Nan::GetFunction(Nan::New(Pattern::constructor)).ToLocalChecked();
+  Local<Function> ctor = Nan::GetFunction(Nan::New(addon_data->pattern_ctor_tpl)).ToLocalChecked();
   Local<Object> instance = Nan::NewInstance(ctor, argc, argv).ToLocalChecked();
 
   info.GetReturnValue().Set(instance);
@@ -2144,9 +2162,10 @@ NAN_METHOD(Context2d::CreatePattern) {
 
 NAN_METHOD(Context2d::CreateLinearGradient) {
   const int argc = 4;
+  AddonData *addon_data = get_data_from_if1(info.This());
   Local<Value> argv[argc] = { info[0], info[1], info[2], info[3] };
+  Local<Function> ctor = Nan::GetFunction(Nan::New(addon_data->gradient_ctor_tpl)).ToLocalChecked();
 
-  Local<Function> ctor = Nan::GetFunction(Nan::New(Gradient::constructor)).ToLocalChecked();
   Local<Object> instance = Nan::NewInstance(ctor, argc, argv).ToLocalChecked();
 
   info.GetReturnValue().Set(instance);
@@ -2154,9 +2173,10 @@ NAN_METHOD(Context2d::CreateLinearGradient) {
 
 NAN_METHOD(Context2d::CreateRadialGradient) {
   const int argc = 6;
+  AddonData *addon_data = get_data_from_if1(info.This());
   Local<Value> argv[argc] = { info[0], info[1], info[2], info[3], info[4], info[5] };
 
-  Local<Function> ctor = Nan::GetFunction(Nan::New(Gradient::constructor)).ToLocalChecked();
+  Local<Function> ctor = Nan::GetFunction(Nan::New(addon_data->gradient_ctor_tpl)).ToLocalChecked();
   Local<Object> instance = Nan::NewInstance(ctor, argc, argv).ToLocalChecked();
 
   info.GetReturnValue().Set(instance);
@@ -2289,8 +2309,9 @@ NAN_METHOD(Context2d::Transform) {
  */
 
 NAN_METHOD(Context2d::GetTransform) {
+  AddonData *addon_data = get_data_from_if1(info.This());
   Context2d *context = Nan::ObjectWrap::Unwrap<Context2d>(info.This());
-  Local<Object> instance = get_current_transform(context);
+  Local<Object> instance = get_current_transform(context, addon_data);
 
   info.GetReturnValue().Set(instance);
 }
@@ -2310,12 +2331,14 @@ NAN_METHOD(Context2d::ResetTransform) {
 
 NAN_METHOD(Context2d::SetTransform) {
   Context2d *context = Nan::ObjectWrap::Unwrap<Context2d>(info.This());
+  AddonData *addon_data = get_data_from_if1(info.This());
   if (info.Length() == 1) {
     Local<Object> mat = Nan::To<Object>(info[0]).ToLocalChecked();
 
     #if NODE_MAJOR_VERSION >= 8
       Local<Context> ctx = Nan::GetCurrentContext();
-      if (!mat->InstanceOf(ctx, _DOMMatrix.Get(Isolate::GetCurrent())).ToChecked()) {
+      Local<Function> _DOMMatrix = Nan::New(addon_data->context2d_dom_matrix);
+      if (!mat->InstanceOf(ctx, _DOMMatrix).ToChecked()) {
         return Nan::ThrowTypeError("Expected DOMMatrix");
       }
     #endif
@@ -2573,16 +2596,17 @@ NAN_GETTER(Context2d::GetFont) {
 NAN_SETTER(Context2d::SetFont) {
   if (!value->IsString()) return;
 
-  Isolate *iso = Isolate::GetCurrent();
   Local<Context> ctx = Nan::GetCurrentContext();
 
   Local<String> str = Nan::To<String>(value).ToLocalChecked();
   if (!str->Length()) return;
 
   const int argc = 1;
+  AddonData *addon_data = get_data_from_if1(info.This());
   Local<Value> argv[argc] = { value };
+  Local<Function> _parseFont = Nan::New(addon_data->context2d_parse_font);
 
-  Local<Value> mparsed = Nan::Call(_parseFont.Get(iso), ctx->Global(), argc, argv).ToLocalChecked();
+  Local<Value> mparsed = Nan::Call(_parseFont, ctx->Global(), argc, argv).ToLocalChecked();
   // parseFont returns undefined for invalid CSS font strings
   if (mparsed->IsUndefined()) return;
   Local<Object> font = Nan::To<Object>(mparsed).ToLocalChecked();
@@ -2612,7 +2636,7 @@ NAN_SETTER(Context2d::SetFont) {
     }
   }
 
-  PangoFontDescription *sys_desc = Canvas::ResolveFontDescription(desc);
+  PangoFontDescription *sys_desc = Canvas::ResolveFontDescription(desc, addon_data);
   pango_font_description_free(desc);
 
   if (size > 0) pango_font_description_set_absolute_size(sys_desc, size * PANGO_SCALE);