+++ src/Image.h
@@ -8,6 +8,7 @@
 #include <nan.h>
 #include <stdint.h> // node < 7 uses libstdc++ on macOS which lacks complete c++11
 #include <v8.h>
+#include "AddonData.h"
 
 #ifdef HAVE_JPEG
 #include <jpeglib.h>
@@ -39,8 +40,8 @@ class Image: public Nan::ObjectWrap {
     char *filename;
     int width, height;
     int naturalWidth, naturalHeight;
-    static Nan::Persistent<v8::FunctionTemplate> constructor;
-    static void Initialize(Nan::ADDON_REGISTER_FUNCTION_ARGS_TYPE target);
+    static const char *ctor_name;
+    static void Initialize(Nan::ADDON_REGISTER_FUNCTION_ARGS_TYPE target, AddonData*);
     static NAN_METHOD(New);
     static NAN_GETTER(GetComplete);
     static NAN_GETTER(GetWidth);